backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic>=2.3.0,<3.0.0
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pandas==2.2.1
scikit-learn==1.4.1.post1
xgboost==2.0.3
shap==0.44.1
joblib==1.3.2
numpy==1.26.4